# myfin-app-batch/src/main/resources/myfinappsswaggerV6withsecurity.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Api",
      "description": "Operations related to accounts"
    },
    {
      "name": "Api",
      "description": "Operations related to markets"
    }
  ],
  "paths": {
    "/api/saveaccounts": {
      "post": {
        "tags": [
          "Api"
        ],
        "summary": "Post accounts to the database",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Post accounts to the database",
        "operationId": "postAccountsToDataBase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AccountHeaderDTO"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "security-controller"
        ],
        "summary": "Post login",
        "description": "Post login",
        "operationId": "authenticateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/markets/{code}/{start}/{end}": {
      "get": {
        "tags": [
          "Api"
        ],
        "summary": "Get market by code",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieves market by its code",
        "operationId": "getMarketByCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MarketDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts": {
      "get": {
        "tags": [
          "Api"
        ],
        "summary": "Get all the account",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieves all the account",
        "operationId": "getAllAccount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountHeaderDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/accountTypes": {
      "get": {
        "tags": [
          "Api"
        ],
        "summary": "Get all the accoun types",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieves all the account types",
        "operationId": "getAccounTypes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/accountSumDate/{start}/{end}": {
      "get": {
        "tags": [
          "Api"
        ],
        "summary": "Get sum of solde by month and year",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieves sum of solde of all accounts by month and year",
        "operationId": "getSumByDate",
        "parameters": [
          {
            "name": "start",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountDateSumDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/accountByTypeSumDate/{accountType}/{start}/{end}": {
      "get": {
        "tags": [
          "Api"
        ],
        "summary": "Get sum of type of solde by month and year",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieves sum of solde of all accounts by month and year",
        "operationId": "getTypeSumByDate",
        "parameters": [
          {
            "name": "accountType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountDateSumDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/account/{id}": {
      "get": {
        "tags": [
          "Api"
        ],
        "summary": "Get account by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieves an account by its id",
        "operationId": "getAccountById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountHeaderDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccountHeaderDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "accountType": {
            "type": "string",
            "enum": [
              "CTO",
              "PEA",
              "PEAPME",
              "ASSURVIE",
              "EPARGNESALARIALE",
              "LIVRETA",
              "LLD",
              "CEP",
              "CPTCOURANT",
              "PRETIMMO"
            ]
          },
          "accountLines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountLineDTO"
            }
          }
        }
      },
      "AccountLineDTO": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "solde": {
            "$ref": "#/components/schemas/AmountOfMoney"
          }
        }
      },
      "AmountOfMoney": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "USD"
            ]
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "usernameOrEmail": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "MarketDTO": {
        "type": "object",
        "properties": {
          "markets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarketDTODetail"
            }
          }
        }
      },
      "MarketDTODetail": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "indicePoint": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "AccountDateSumDTO": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "sum": {
            "type": "number"
          },
          "cuurency": {
            "type": "string",
            "enum": [
              "EUR",
              "USD"
            ]
          }
        }
      }
    },
     "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}